+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## [Unreleased]
 
-- Switch packaging and project workflows to poetry
-- Drop support for python 3.7
-- Drop support for Exasol 6.x
-- Drop support for Exasol 7.0.x
+## [0.26.0] - 2024-07-04
+
+- Added dbapi2 compliant driver interface `exasol.driver.websocket` ontop of pyexasol
+
+    ⚠️ Note:
+
+    This driver facade should only be used if one is certain that using the dbapi2 is the right solution for their scenario, taking all implications into account. For more details on why and who should avoid using dbapi2, please refer to the [DBAPI2 compatibility section](/docs/DBAPI_COMPAT.md) in our documentation.
+
+- Droped support for python 3.7
+- Droped support for Exasol 6.x
+- Droped support for Exasol 7.0.x
+- Relocked dependencies (Internal)
+- Switched packaging and project workflow to poetry (internal)
 
 ## [0.25.2] - 2023-03-14